requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email-validator
python-dotenv
# psycopg2-binary  <-- Removed this